kept only buisness logic in hook - acc to review
