+++ packages/twenty-front/src/modules/object-record/record-index/components/RecordIndexPageKanbanAddButton.tsx
@@ -1,19 +1,29 @@
+import { CoreObjectNameSingular } from '@/object-metadata/types/CoreObjectNameSingular';
+import { useRecordBoardStates } from '@/object-record/record-board/hooks/internal/useRecordBoardStates';
+import { RecordBoardColumnDefinition } from '@/object-record/record-board/types/RecordBoardColumnDefinition';
+import { RecordIndexPageKanbanAddMenuItem } from '@/object-record/record-index/components/RecordIndexPageKanbanAddMenuItem';
+import { useRecordIndexPageKanbanAddButton } from '@/object-record/record-index/hooks/useRecordIndexPageKanbanAddButton';
+import { SingleEntitySelect } from '@/object-record/relation-picker/components/SingleEntitySelect';
+import { useEntitySelectSearch } from '@/object-record/relation-picker/hooks/useEntitySelectSearch';
+import { EntityForSelect } from '@/object-record/relation-picker/types/EntityForSelect';
+import { RelationPickerHotkeyScope } from '@/object-record/relation-picker/types/RelationPickerHotkeyScope';
+import { IconButton } from '@/ui/input/button/components/IconButton';
 import { Dropdown } from '@/ui/layout/dropdown/components/Dropdown';
 import { DropdownMenu } from '@/ui/layout/dropdown/components/DropdownMenu';
 import { DropdownMenuItemsContainer } from '@/ui/layout/dropdown/components/DropdownMenuItemsContainer';
+import { useDropdown } from '@/ui/layout/dropdown/hooks/useDropdown';
+import { usePreviousHotkeyScope } from '@/ui/utilities/hotkey/hooks/usePreviousHotkeyScope';
 import styled from '@emotion/styled';
-import { useState } from 'react';
-
-import { SingleEntitySelect } from '@/object-record/relation-picker/components/SingleEntitySelect';
-
-import { CoreObjectNameSingular } from '@/object-metadata/types/CoreObjectNameSingular';
-import { RecordIndexPageKanbanAddMenuItem } from '@/object-record/record-index/components/RecordIndexPageKanbanAddMenuItem';
-import { useRecordIndexPageKanbanAddButton } from '@/object-record/record-index/hooks/useRecordIndexPageKanbanAddButton';
-import { IconButton } from '@/ui/input/button/components/IconButton';
-import { IconPlus } from 'twenty-ui';
+import { useCallback, useState } from 'react';
+import { useRecoilValue } from 'recoil';
+import { IconPlus, isDefined } from 'twenty-ui';
 
 const StyledDropdownMenuItemsContainer = styled(DropdownMenuItemsContainer)`
   width: 100%;
+`;
+
+const StyledDropDownMenu = styled(DropdownMenu)`
+  width: 200px;
 `;
 
 type RecordIndexPageKanbanAddButtonProps = {
@@ -25,22 +35,78 @@ export const RecordIndexPageKanbanAddButton = ({
   recordIndexId,
   objectNamePlural,
 }: RecordIndexPageKanbanAddButtonProps) => {
+  const dropdownId = `record-index-page-add-button-dropdown`;
   const [isSelectingCompany, setIsSelectingCompany] = useState(false);
+  const [selectedColumnDefinition, setSelectedColumnDefinition] =
+    useState<RecordBoardColumnDefinition>();
+
+  const { columnIdsState } = useRecordBoardStates(recordIndexId);
+  const columnIds = useRecoilValue(columnIdsState);
 
   const {
-    dropdownId,
-    columnIds,
+    setHotkeyScopeAndMemorizePreviousScope,
+    goBackToPreviousHotkeyScope,
+  } = usePreviousHotkeyScope();
+  const { resetSearchFilter } = useEntitySelectSearch({
+    relationPickerScopeId: 'relation-picker',
+  });
+
+  const { closeDropdown } = useDropdown(dropdownId);
+
+  const {
     selectFieldMetadataItem,
     isOpportunity,
-    handleOpportunityClick,
-    handleNonOpportunityClick,
-    handleEntitySelect,
-    handleCancel,
+    createOpportunity,
+    createRecordWithoutCompany,
   } = useRecordIndexPageKanbanAddButton({
-    recordIndexId,
     objectNamePlural,
-    setIsSelectingCompany,
   });
+
+  const handleItemClick = useCallback(
+    (columnDefinition: RecordBoardColumnDefinition) => {
+      if (isOpportunity) {
+        setIsSelectingCompany(true);
+        setSelectedColumnDefinition(columnDefinition);
+        setHotkeyScopeAndMemorizePreviousScope(
+          RelationPickerHotkeyScope.RelationPicker,
+        );
+      } else {
+        createRecordWithoutCompany(columnDefinition);
+        closeDropdown();
+      }
+    },
+    [
+      isOpportunity,
+      createRecordWithoutCompany,
+      setHotkeyScopeAndMemorizePreviousScope,
+      closeDropdown,
+    ],
+  );
+
+  const handleEntitySelect = useCallback(
+    (company?: EntityForSelect) => {
+      setIsSelectingCompany(false);
+      goBackToPreviousHotkeyScope();
+      resetSearchFilter();
+      if (isDefined(company) && isDefined(selectedColumnDefinition)) {
+        createOpportunity(company, selectedColumnDefinition);
+      }
+      closeDropdown();
+    },
+    [
+      createOpportunity,
+      goBackToPreviousHotkeyScope,
+      resetSearchFilter,
+      selectedColumnDefinition,
+      closeDropdown,
+    ],
+  );
+
+  const handleCancel = useCallback(() => {
+    resetSearchFilter();
+    goBackToPreviousHotkeyScope();
+    setIsSelectingCompany(false);
+  }, [goBackToPreviousHotkeyScope, resetSearchFilter]);
 
   if (!selectFieldMetadataItem) {
     return null;
@@ -62,7 +128,7 @@ export const RecordIndexPageKanbanAddButton = ({
       }
       dropdownId={dropdownId}
       dropdownComponents={
-        <DropdownMenu style={{ width: '200px' }}>
+        <StyledDropDownMenu>
           {isOpportunity && isSelectingCompany ? (
             <SingleEntitySelect
               disableBackgroundBlur
@@ -79,16 +145,12 @@ export const RecordIndexPageKanbanAddButton = ({
                   key={columnId}
                   columnId={columnId}
                   recordIndexId={recordIndexId}
-                  onItemClick={
-                    isOpportunity
-                      ? handleOpportunityClick
-                      : handleNonOpportunityClick
-                  }
+                  onItemClick={handleItemClick}
                 />
               ))}
             </StyledDropdownMenuItemsContainer>
           )}
-        </DropdownMenu>
+        </StyledDropDownMenu>
       }
       dropdownHotkeyScope={{ scope: dropdownId }}
     />
